New SHT_RTC

+++ src/A_Core/Interfaces.h
@@ -9,7 +9,7 @@ class IObserver
 {
 public:
     virtual ~IObserver() {}
-    virtual void onNotify(EventFlag::Type flag, const float value) {}
+    virtual void onNotify(EventFlag::Type flag, const uint16_t value) {}
     virtual void onNotify(EventFlag::Type flag, const uint8_t value) {}
     virtual void onNotify(EventFlag::Type flag, const uint32_t value) {}
     //virtual void onNotify(EventFlag::Type flag, const SystemConfig& value) {} // Config Load & Save
